fix travis commit

+++ src/dpdk/lib/librte_eal/common/include/rte_service.h
@@ -337,7 +337,7 @@ int32_t rte_service_set_stats_enable(uint32_t id, int32_t enable);
 int32_t rte_service_lcore_list(uint32_t array[], uint32_t n);
 
 /**
- * Get the number of services running on the supplied lcore.
+ * Get the numer of services running on the supplied lcore.
  *
  * @param lcore Id of the service core.
  * @retval >=0 Number of services registered to this core.
@@ -372,7 +372,7 @@ int32_t rte_service_dump(FILE *f, uint32_t id);
  *         -EINVAL Invalid id, attr_id or attr_value was NULL.
  */
 int32_t rte_service_attr_get(uint32_t id, uint32_t attr_id,
-		uint64_t *attr_value);
+		uint32_t *attr_value);
 
 /**
  * Reset all attribute values of a service.
